consolidating prereqs logic
