+++ app/routes/app+/market_.$address.request.$requestId.tsx
@@ -18,9 +18,7 @@ import { ParsedMarkdown } from "~/features/markdown-editor/markdown.client";
 import { ProjectBadges } from "~/features/project-badges";
 import { WalletGuardedButtonLink } from "~/features/wallet-guarded-button-link";
 import { useHasPerformed } from "~/hooks/use-has-performed";
-import { useReputationTokenBalance } from "~/hooks/use-reputation-token-balance";
 import { useReviewSignals } from "~/hooks/use-review-signals";
-import { useTokenBalance } from "~/hooks/use-token-balance";
 import { useOptionalUser } from "~/hooks/use-user";
 import { findProjectsBySlug } from "~/services/projects.server";
 import { countReviews } from "~/services/review-service.server";
@@ -30,6 +28,7 @@ import { REPUTATION_REWARD_POOL } from "~/utils/constants";
 import { dateHasPassed } from "~/utils/date";
 import { claimToReviewDeadline, fromTokenAmount } from "~/utils/helpers";
 import * as DOMPurify from "dompurify";
+import { usePrereqs } from "~/hooks/use-prereqs";
 
 const paramsSchema = z.object({ address: z.string(), requestId: z.string() });
 export const loader = async ({ params }: DataFunctionArgs) => {
@@ -77,31 +76,21 @@ export default function ServiceRequest() {
     action: "HAS_SUBMITTED",
   });
 
-  const maintainerBadgeTokenBalance = useTokenBalance({
-    tokenAddress: laborMarket.configuration.maintainerBadge.token as `0x${string}`,
-    tokenId: laborMarket.configuration.maintainerBadge.tokenId,
-  });
-
   const reviewSignal = useReviewSignals({
     laborMarketAddress: serviceRequest.laborMarketAddress as `0x${string}`,
     serviceRequestId: serviceRequest.id,
   });
 
-  const reputationBalance = useReputationTokenBalance();
-
   const user = useOptionalUser();
   const userSignedIn = !!user;
 
   const showSubmit = hasClaimedToSubmit && !hasSubmitted;
-  const canClaimToSubmit =
-    reputationBalance?.gte(laborMarket.configuration.reputationParams.submitMin) &&
-    reputationBalance?.lte(laborMarket.configuration.reputationParams.submitMax);
   const showClaimToSubmit = !hasClaimedToSubmit && !hasSubmitted && !claimDeadlinePassed;
   const showClaimToReview =
     reviewSignal?.remainder.eq(0) && // Must not have any remaining reviews left (or initial of 0)
     !claimToReviewDeadlinePassed;
 
-  const canClaimToReview = maintainerBadgeTokenBalance?.gt(0);
+  const { canReview, canSubmit } = usePrereqs({ laborMarket });
 
   return (
     <Container className="pt-7 pb-16 px-10">
@@ -118,7 +107,7 @@ export default function ServiceRequest() {
             <WalletGuardedButtonLink
               buttonText="Claim to Review"
               link={`/app/market/${laborMarket.address}/request/${serviceRequest.id}/review`}
-              disabled={userSignedIn && !canClaimToReview}
+              disabled={userSignedIn && !canReview}
               disabledTooltip="Check for Prerequisites"
               variant="cancel"
               size="lg"
@@ -128,7 +117,7 @@ export default function ServiceRequest() {
             <WalletGuardedButtonLink
               buttonText="Claim to Submit"
               link={`/app/market/${laborMarket.address}/request/${serviceRequest.id}/claim`}
-              disabled={userSignedIn && !canClaimToSubmit}
+              disabled={userSignedIn && !canSubmit}
               disabledTooltip="Check for Prerequisites"
               size="lg"
             />
